Почистил код
